+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Bigsby's Hand 2024 Summoning.
 - Dazzling Footwork enchantment. (Requires D&D system 4.0.4 or 4.1.x).
+- Some College of Dance bard features.
 
 # 6.0.10